# perf: batch cache processing