This patch is recommended by AI, and it reduces the time from

Loading index from cache: 1383.408935546875 ms
plugin:omnisearch:104 Indexing total time: 1385.22607421875 ms

to

Loading index from cache: 1216.864013671875 ms
plugin:omnisearch:106 Indexing total time: 1218.73291015625 ms

May need more testing though