@qazxcdswe123 2025-01-01
For the second patch:

Loading index from cache: 1165.33203125 ms
plugin:omnisearch:106 Indexing total time: 1167.251953125 ms