@scambier 2025-01-03
I've benchmarked the changes myself by repeatedly running the "Reload app without saving" command to have cold runs, and the results are at best inconclusive.

The 2 last runs I've done :
- Without PDFs
  - Original code: 2108ms 
  - Your changes: 2126ms
- With PDFS
  - Original code: 10791ms
  - Your changes: 10766ms 

I've also noted that `await this.plugin.database.minisearch.limit(1).first()` is almost consistently slower than `(await this.plugin.database.minisearch.toArray())[0]`. That's a bit surprising, but Omnisearch's IndexedDb is maybe incorrectly configured 🤷‍♂️